experiments/requirements.txt
copernicusmarine
clouddrift
dask
equinox
h5py
hydra-zen
jax
kymatio
lightning
numpy
optax
pastax==0.0.6
requests
torchpip
torchvision
tqdm
xarray